+++ apps/api/src/task/dto/add-task-label.dto.ts
@@ -1,0 +1,7 @@
+import { IsNotEmpty, IsUUID } from 'class-validator';
+
+export class AddTaskLabelDto {
+  @IsUUID('7')
+  @IsNotEmpty()
+  labelId!: string;
+}